Rename `_EquationGenerator` to `ConstraintGenerator`
First, I wanted to make the class public because someone else might
want to use it in a different implementation of type inference and
making the class public makes it easier to do that.

Second, constraint generator is a better name overall because later on
I might implement constraints other than equality (e.g. for bounded
polymorphism).