Fix a bug
This bug was causing `lowered.Block` nodes to compare the types of the
containers of their `body` attrs when checking if they are equivalent.
This meant that a case where 2 `lowered.Block`s were identical except
that 1 used a `list` for its body and the other used a `tuple` was
marked wrong.